pyproject.toml
[tool.poetry]
name = "dbsynchronization"
version = "0.1.0"
description = ""
authors = ["Rohan Dsouza <r.dsouza@somaiya.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymysql = "^1.1.1"
psycopg2 = "^2.9.9"
pyyaml = "^6.0.1"
flask = "^3.0.3"
cryptography = "^43.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
